+++ ControlAcceso.Tests/Data/UsersDbContextTests.cs
@@ -1,7 +1,6 @@
 ﻿using System.Data.Common;
 using ControlAcceso.Data.Model;
 using ControlAcceso.Data.Users;
-using ControlAcceso.Data.Roles;
 using ControlAcceso.Services.DBService;
 using Moq;
 
@@ -239,32 +238,5 @@ namespace ControlAcceso.Tests.Data
                 )
             ), Times.Once);
         }
-
-        [Fact]
-        public void RoleExists_ShouldReturnTrue_WhenRoleExists()
-        {
-            // Arrange
-            int roleId = 1;
-            var expectedResult = new List<Dictionary<string, object>>
-            {
-                new Dictionary<string, object> { { "Exists", 1 } }
-            };
-
-            // Mock
-            _dbServiceMock.Setup(x => x.ExecuteReader(
-                It.Is<string>(query => query.Contains("SELECT 1")), 
-                It.Is<Dictionary<string, object>>(parameters => (int)parameters["@RoleId"] == roleId)))
-                .Returns(expectedResult);  
-
-            var context = new RolesDbContext(_dbServiceMock.Object);
-
-            // Act
-            var result = context.RoleExists(roleId);
-
-            // Assert
-            Assert.True(result); 
-        }
-
-
     }
 }
